The commit
892deb766f02fc4e9d6ce0c8c1ef55eca62476df

contains the version with the (correct) no cspool in filters.lib

In the commit 50766896885b7f1c0a6ae7c584de13b65da05877
we introduced the revision to the rad processing in filters.lib,
but somehow we used the wrong filters.lib file that included the cspool.

The solution was to take the filters.lib from

892deb766f02fc4e9d6ce0c8c1ef55eca62476df

and use the function "filterslib_get_rcvars{}" from

50766896885b7f1c0a6ae7c584de13b65da05877.

The revision os only the addition of the statement
with the following piece: (filterslib(doradinfounzretry) == 1).